add temp and city display
